requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
futures==3.0.5
goslate==1.5.4
idna==3.10
lxml==5.4.0
mysql-connector-python==9.3.0
pyTelegramBotAPI==4.27.0
python-docx==1.2.0
python-dotenv==1.1.0
requests==2.32.4
typing_extensions==4.14.0
urllib3==2.5.0
